requirements.txt
matplotlib==3.8.2
numpy==1.24.3
pandas==2.2.0
yfinance==0.2.36
pyqt==5.15.9
